+++ src/app/api/admin/usage/route.ts
@@ -1,23 +1,20 @@
 import { NextResponse } from 'next/server';
 import { getUsageStats } from '@/lib/api';
+import { ENV } from '@/types/envSchema';
 
-export function GET(request: Request) {
-  // In production, add authentication check here
+export async function GET(request: Request): Promise<Response | void> {
   const authHeader = request.headers.get('authorization');
-  
-  // Simple auth check - replace with proper auth in production
+
   if (!authHeader || !authHeader.startsWith('Bearer ')) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
-  
-  // Get token from header
+
   const token = authHeader.split(' ')[1];
-  
-  // Check if token is valid (use a secure method in production)
-  if (token !== process.env.ADMIN_API_KEY) {
+
+  if (token !== ENV.ADMIN_API_KEY) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
-  
-  // Return usage stats
-  return NextResponse.json(getUsageStats());
+
+  const stats = await getUsageStats();
+  return NextResponse.json(stats);
 }
