+++ src/app/api/chat/homeworkhelper/route.ts
@@ -5,9 +5,9 @@ import { trimConversationHistory } from '@/lib/tokenManagement';
 
 export async function POST(request: Request) {
   try {
-    const body = await request.json() as OpenAIRequestBody;
+    const body = (await request.json()) as OpenAIRequestBody;
     
-    // Validate request
+    // Validate the chat request
     const validation = validateChatRequest(body);
     if (!validation.isValid) {
       return NextResponse.json(
@@ -15,32 +15,30 @@ export async function POST(request: Request) {
         { status: 400 }
       );
     }
-
-    // Trim conversation history for token management
+    
+    // Trim conversation history for better token management
     const trimmedMessages = trimConversationHistory(body.messages, 4000, 4);
     
-    // Add system message for homework helper mode
+    // Add a system prompt for homework helper mode
     const messages = addSystemPrompt(trimmedMessages, 'homeworkhelper');
     
-    // Get last user message to determine homework subject and complexity
-    const lastUserMsg = [...messages]
-      .reverse()
-      .find(m => m.role === 'user')?.content || '';
+    // Retrieve the last user message to decide on subject complexity
+    const lastUserMsg = [...messages].reverse().find(m => m.role === 'user')?.content || '';
       
-    // Select model based on subject area and complexity
+    // Select model based on advanced STEM keywords in the last user message
     const isAdvancedMath = /calculus|differential equations|linear algebra|statistics|probability|vectors/i.test(lastUserMsg);
     const isAdvancedScience = /physics|chemistry|biochemistry|quantum|thermodynamics|organic chemistry/i.test(lastUserMsg);
     
-    // Use more capable model for advanced STEM subjects
+    // Use a more capable model for advanced STEM topics
     const model = (isAdvancedMath || isAdvancedScience) ? "gpt-4o" : "gpt-3.5-turbo";
-
-    // Process with OpenAI
+    
+    // Process the chat request with OpenAI
     return processChatRequest(messages, {
       max_tokens: body.max_tokens || 1000,
-      temperature: 0.5, // More precise for educational content
+      temperature: 0.5, // Use a lower temperature for more precise educational content
       model: model,
       mode: 'homeworkhelper',
-      useCache: false // Don't cache homework help as it should be personalized
+      useCache: false // Disable caching to ensure personalized homework help
     });
   } catch (error) {
     console.error('Error in homework helper mode:', error);
