+++ src/app/api/chat/reason/route.ts
@@ -1,12 +1,11 @@
-
 import { NextResponse } from 'next/server';
 import { OpenAIRequestBody } from '@/types/openai';
 import { validateChatRequest, addSystemPrompt, processChatRequest } from '@/lib/api';
-import { trimConversationHistory, selectAppropriateModel } from '@/lib/tokenManagement';
+import { trimConversationHistory } from '@/lib/tokenManagement';
 
 export async function POST(request: Request) {
   try {
-    const body = await request.json() as OpenAIRequestBody;
+    const body = (await request.json()) as OpenAIRequestBody;
     
     // Validate request
     const validation = validateChatRequest(body);
@@ -17,22 +16,22 @@ export async function POST(request: Request) {
       );
     }
 
-    // Trim conversation history for token management
-    const trimmedMessages = trimConversationHistory(body.messages, 6000, 6); // Keep more context for reasoning
+    // Trim conversation history for token management (keeping more context for reasoning)
+    const trimmedMessages = trimConversationHistory(body.messages, 6000, 6);
     
     // Add system message for reasoning mode
     const messages = addSystemPrompt(trimmedMessages, 'reason');
     
-    // Always use GPT-4o for reasoning tasks which require more advanced capabilities
+    // Always use GPT-4o for reasoning tasks
     const model = "gpt-4o";
 
-    // Process with OpenAI
+    // Process the request with OpenAI
     return processChatRequest(messages, {
-      max_tokens: body.max_tokens || 1200, // Allow longer responses for reasoning
-      temperature: 0.7, // Balanced for reasoning
+      max_tokens: body.max_tokens || 1200, // Longer responses for reasoning tasks
+      temperature: 0.7, // Balanced temperature for reasoning
       model: model,
       mode: 'reason',
-      useCache: false // Reasoning should be fresh each time
+      useCache: false // Reasoning should be processed fresh each time
     });
   } catch (error) {
     console.error('Error in reasoning mode:', error);
