+++ src/app/create/topics/page.tsx
@@ -1,7 +1,13 @@
 "use client";
- 
- import SubjectsTopicsPage from "@/components/topics";
- 
- export default function TopicsPage() {
-   return <SubjectsTopicsPage />;
- }
+
+import SubjectsTopicsPage from "@/components/topics";
+import Loader from "@/components/ui/loader";
+import { Suspense } from "react";
+
+export default function TopicsPage() {
+  return (
+    <Suspense fallback={<Loader />}>
+      <SubjectsTopicsPage />;
+    </Suspense>
+  );
+}
