+++ src/components/ui/loader.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function Loader() {
+  return (
+    <div className="flex flex-row gap-2">
+      <div className="w-4 h-4 rounded-full bg-blue-700 animate-bounce"></div>
+      <div className="w-4 h-4 rounded-full bg-blue-700 animate-bounce [animation-delay:-.3s]"></div>
+      <div className="w-4 h-4 rounded-full bg-blue-700 animate-bounce [animation-delay:-.5s]"></div>
+    </div>
+  );
+}
